% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtracks.R
\name{calculate_geographic_overlap_and_segregation}
\alias{calculate_geographic_overlap_and_segregation}
\title{Analysis of geographic overlap and geographic segregation between two groups of xtracks}
\usage{
calculate_geographic_overlap_and_segregation(
  xtrack_list_1,
  xtrack_list_2,
  cell_size_m = 10
)
}
\arguments{
\item{xtrack_list_1}{a list of xtracks representing 'group 1'}

\item{xtrack_list_2}{a list of xtracks representing 'group 1'.}

\item{cell_size_m}{The resolution of the raster analysis -- i.e. the height and width of each cell in the raster representation of the landscape, in meters.}
}
\value{
A list presenting analysis results
\describe{
  \item{square_meters_visited_1}{square meters of unique landscape areas visited by all xtracks in group 1}
  \item{square_meters_visited_2}{square meters of unique landscape areas visited by all xtracks in group 2}
  \item{square_meters_visited_1_or_2}{square meters of unique landscape areas visited by 1 or 2 (geographic union)}
  \item{square_meters_visited_by_1_and_2}{square meters of unique landscape areas visited by 1 and 2 (geographic intersection)}
  \item{percent_of_what_1_visited_that_was_visited_by_2}{percent of land visited by group 1 that was visited by group 2}
  \item{percent_of_what_2_visited_that_was_visited_by_1}{percent of land visited by group 2 that was visited by group 1}
  \item{percent_of_what_was_visited_by_1_or_2_that_was_visited_by_1}{percent of unique land visited by 1 or 2 that was visited by 1}
  \item{percent_of_what_was_visited_by_1_or_2_that_was_visited_by_2}{percent of unique land visited by 1 or 2 that was visited by 2}
  \item{percent_of_what_was_visited_by_1_or_2_that_was_visited_by_1_and_2}{percent of unique land visited by 1 or 2 that was visited by both 1 and 2.}}
}
\description{
This analysis calculates measures of geographic overlap and geographic segregation
between all the xtracks in one group and all the xtracks in another group. These groups could be gender groups (as done
in Wood et al. 2021), or they could represent age groups, family groups, gender X age groups,
all could be potentially interesting. This analysis should be done with two lists of xtrack objects, with each xtrack representing
one day of travel. The lists can be of any length, but for an apples-to-apples empirical comparison, should be approx. equal in length.
}
\examples{
data(d1,d2,d3,d4,d5,d6,d7,d8)
xt_1 <- xtrack(lat=d1$lat, lon=d1$lon, elevation_m=d1$elevation_m, in_camp=d1$in_camp, unix_time=d1$unix_time, distance_from_camp_m=d1$distance_from_camp_m, utm_epsg=32736)
xt_2 <- xtrack(lat=d2$lat, lon=d2$lon, elevation_m=d2$elevation_m, in_camp=d2$in_camp, unix_time=d2$unix_time, distance_from_camp_m=d2$distance_from_camp_m, utm_epsg=32736)
xt_3 <- xtrack(lat=d3$lat, lon=d3$lon, elevation_m=d3$elevation_m, in_camp=d3$in_camp, unix_time=d3$unix_time, distance_from_camp_m=d3$distance_from_camp_m, utm_epsg=32736)
xt_4 <- xtrack(lat=d4$lat, lon=d4$lon, elevation_m=d4$elevation_m, in_camp=d4$in_camp, unix_time=d4$unix_time, distance_from_camp_m=d4$distance_from_camp_m, utm_epsg=32736)
xt_5 <- xtrack(lat=d5$lat, lon=d5$lon, elevation_m=d5$elevation_m, in_camp=d5$in_camp, unix_time=d5$unix_time, distance_from_camp_m=d5$distance_from_camp_m, utm_epsg=32736)
xt_6 <- xtrack(lat=d6$lat, lon=d6$lon, elevation_m=d6$elevation_m, in_camp=d6$in_camp, unix_time=d6$unix_time, distance_from_camp_m=d6$distance_from_camp_m, utm_epsg=32736)
xt_7 <- xtrack(lat=d7$lat, lon=d7$lon, elevation_m=d7$elevation_m, in_camp=d7$in_camp, unix_time=d7$unix_time, distance_from_camp_m=d7$distance_from_camp_m, utm_epsg=32736)
xt_8 <- xtrack(lat=d8$lat, lon=d8$lon, elevation_m=d8$elevation_m, in_camp=d8$in_camp, unix_time=d8$unix_time, distance_from_camp_m=d8$distance_from_camp_m, utm_epsg=32736)
list_1 <- list(xt_1, xt_2, xt_3, xt_4)
list_2 <- list(xt_5, xt_6, xt_7, xt_8)
geo_seg_results <- calculate_geographic_overlap_and_segregation(list_1, list_2, cell_size_m = 10)
}
\seealso{
For example of geographic segregation and overlap analysis see figure 4 in Wood et al. 2021 publication see \url{https://www.nature.com/articles/s41562-020-01002-7}
}
