% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d7}
\alias{d7}
\title{Individual GPS trajectory dataset}
\format{
A data frame with one row per trackpoint and 6 variables per trackpoint:
\describe{
  \item{lat}{latitude, in decimal degrees, WGS 84}
  \item{lon}{longitude, in decimal degrees, WGS 84}
  \item{elevation_m}{elevation in meters}
  \item{in_camp}{binary flag to indicate if the individual is in the camp / residential area, or out of the camp / residential area}
  \item{unix_time}{the GPS timestamp, encoded as the number of seconds since January 1st, 1970 UTC timezone}
  \item{distance_from_camp_m}{the distance of the individual from the centroid of the camp / residential area}
}
}
\usage{
d7
}
\description{
Demonstration GPS data formatted in manner to facilitate the construction of an xtrack object and demonstration of package xtracks features.
For privacy concerns, the raw spatial data have been shifted such that this track will appear to be in the center of Lake Eyasi when mapped.
}
\keyword{datasets}
